+++ base64.h
@@ -17,7 +17,7 @@ std::string base64_encode_pem (std::string const& s);
 std::string base64_encode_mime(std::string const& s);
 
 std::string base64_decode(std::string const& s, bool remove_linebreaks = false);
-std::string base64_encode(unsigned char const*, unsigned int len, bool url = false);
+std::string base64_encode(unsigned char const*, size_t len, bool url = false);
 
 #if __cplusplus >= 201703L
 //
